# src/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Swarmail",
    "version": "0.1.0",
    "description": "The fastest AI-native SMTP mail mock. Accepts any SMTP traffic on :1025, answers POP3 (RFC 1939 core + TOP/CAPA) on :1110 with the inbox as the maildrop, and exposes captured email through this API on :8025 — including MailHog/Mailpit-compatible read endpoints. Every accepted message is queryable immediately — there is no async gap where mail can be lost."
  },
  "paths": {
    "/healthz": {
      "get": {
        "summary": "Liveness/uptime",
        "responses": { "200": { "description": "Server status" } }
      }
    },
    "/metrics": {
      "get": { "summary": "Prometheus metrics", "responses": { "200": { "description": "text/plain metrics" } } }
    },
    "/openapi.json": { "get": { "summary": "This document", "responses": { "200": { "description": "OK" } } } },
    "/llms.txt": { "get": { "summary": "LLM-oriented quick reference", "responses": { "200": { "description": "text/plain" } } } },
    "/mcp": {
      "post": {
        "summary": "Model Context Protocol endpoint (JSON-RPC 2.0)",
        "description": "initialize, tools/list, tools/call. The stdio transport is available via `swarmail mcp`.",
        "responses": { "200": { "description": "JSON-RPC response" }, "202": { "description": "Notification accepted" } }
      }
    },
    "/api/v1/inboxes": {
      "get": {
        "summary": "List inboxes with message counts",
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/InboxSummary" } } } } } }
      }
    },
    "/api/v1/inboxes/{inbox}/messages": {
      "get": {
        "summary": "List messages in an inbox (newest first)",
        "parameters": [
          { "name": "inbox", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "to", "in": "query", "schema": { "type": "string" }, "description": "Substring match on any recipient (to/cc)" },
          { "name": "from", "in": "query", "schema": { "type": "string" } },
          { "name": "subject", "in": "query", "schema": { "type": "string" } },
          { "name": "since_ms", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EmailList" } } } } }
      },
      "delete": {
        "summary": "Clear an inbox",
        "parameters": [{ "name": "inbox", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Removed count" } }
      }
    },
    "/api/v1/inboxes/{inbox}/count": {
      "get": {
        "summary": "Count messages matching a filter",
        "parameters": [
          { "name": "inbox", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "to", "in": "query", "schema": { "type": "string" } },
          { "name": "from", "in": "query", "schema": { "type": "string" } },
          { "name": "subject", "in": "query", "schema": { "type": "string" } },
          { "name": "since_ms", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "{ inbox, count }" } }
      }
    },
    "/api/v1/inboxes/{inbox}/await": {
      "get": {
        "summary": "Long-poll until `count` messages match the filter (push-based wait, no sleep-polling)",
        "parameters": [
          { "name": "inbox", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "to", "in": "query", "schema": { "type": "string" } },
          { "name": "subject", "in": "query", "schema": { "type": "string" } },
          { "name": "count", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "timeout_ms", "in": "query", "schema": { "type": "integer", "default": 5000 } }
        ],
        "responses": {
          "200": { "description": "Matched", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AwaitResponse" } } } },
          "408": { "description": "Timed out" }
        }
      }
    },
    "/api/v1/inboxes/{inbox}/assert": {
      "get": {
        "summary": "Assert an exact match count (200 when equal, 409 with expected/matched otherwise)",
        "parameters": [
          { "name": "inbox", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "to", "in": "query", "schema": { "type": "string" } },
          { "name": "count", "in": "query", "schema": { "type": "integer", "default": 1 } }
        ],
        "responses": { "200": { "description": "Assertion held" }, "409": { "description": "Assertion failed" } }
      }
    },
    "/api/v1/inboxes/{inbox}/threads": {
      "get": {
        "summary": "Thread view of an inbox (newest thread first)",
        "description": "Mails are grouped into conversations by their References/In-Reply-To/Message-ID chains, falling back to the normalized subject (Re:/Fwd: markers stripped, case-folded) when a chain cannot resolve. Each thread lists its mails oldest first. Capped at 200 mails.",
        "parameters": [{ "name": "inbox", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Thread" } } } } } }
      }
    },
    "/api/v1/inboxes/{inbox}/threads/{key}": {
      "get": {
        "summary": "One thread by its key (the conversation, oldest first)",
        "parameters": [
          { "name": "inbox", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "key", "in": "path", "required": true, "schema": { "type": "string" }, "description": "The thread's stable key, as returned by the threads list" }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Thread" } } } }, "404": { "description": "Not found" } }
      }
    },
    "/api/v1/inboxes/{inbox}/feed": {
      "get": {
        "summary": "Live inbox feed (server-sent events)",
        "description": "One `threads` event per change, carrying the full thread view (the GET /threads shape) as JSON. The first event is the snapshot; a watcher that falls behind is repaired with a recomputed view, never an error. Subscribes before the first scan, so mail landing while the client connects is pushed, not missed.",
        "parameters": [{ "name": "inbox", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "text/event-stream" } }
      }
    },
    "/api/v1/inboxes/{inbox}/seed": {
      "post": {
        "summary": "Inject a synthetic email (full parse/extract pipeline runs)",
        "parameters": [{ "name": "inbox", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SeedRequest" } } } },
        "responses": { "200": { "description": "The stored email" } }
      }
    },
    "/api/v1/messages": {
      "get": {
        "summary": "MailHog/Mailpit list shape over the whole store (Mailpit's summary envelope, newest first)",
        "description": "The Mailpit `/api/v1/messages` shape: {total, unread, count, messages_count, messages_unread_count, start, tags, messages}. No query names a mailbox: the whole store is one global mailbox, as both tools expose it. `?inbox=` scopes to a single swarmail inbox; `?limit=`/`?start=` page.",
        "parameters": [
          { "name": "inbox", "in": "query", "schema": { "type": "string" }, "description": "Scope to one swarmail inbox" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "start", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": { "200": { "description": "Mailpit MessageSummary envelope" } }
      },
      "delete": {
        "summary": "Delete the named ids (Mailpit) or wipe every inbox (MailHog)",
        "description": "Body `{\"ids\":[...]}` deletes exactly those messages; an absent, empty or unparseable body wipes everything, like upstream's decoders. Replies `{\"removed\": n}`.",
        "responses": { "200": { "description": "{\"removed\": n}" } }
      }
    },
    "/api/v1/message/{id}": {
      "get": {
        "summary": "Mailpit's full message shape",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Mailpit Message" }, "404": { "description": "Not found" } }
      }
    },
    "/api/v1/message/{id}/plain": {
      "get": {
        "summary": "The extracted text part as text/plain (Mailpit + MailHog)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "text/plain; charset=utf-8" }, "404": { "description": "Not found" } }
      }
    },
    "/api/v1/message/{id}/raw": {
      "get": {
        "summary": "The full RFC 5322 source (MailHog's raw shape)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "message/rfc822" }, "404": { "description": "Not found" } }
      }
    },
    "/api/v1/message/{id}/headers": {
      "get": {
        "summary": "The synthesized header map (MailHog's Content.Headers)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Header name to array of values, including Received and Return-Path stamps" }, "404": { "description": "Not found" } }
      }
    },
    "/api/v1/messages/{id}/plain": {
      "get": {
        "summary": "The extracted text part as text/plain (the plural path MailHog clients use)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "text/plain; charset=utf-8" }, "404": { "description": "Not found" } }
      }
    },
    "/api/v1/messages/{id}/download": {
      "get": {
        "summary": "The message as an .eml attachment download",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "message/rfc822, Content-Disposition: attachment; filename=\"<id>.eml\"" }, "404": { "description": "Not found" } }
      }
    },
    "/api/v1/search": {
      "get": {
        "summary": "Mailpit search envelope over the store",
        "parameters": [
          { "name": "query", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "kind", "in": "query", "schema": { "type": "string", "enum": ["from", "to", "subject", "containing"], "default": "containing" } },
          { "name": "inbox", "in": "query", "schema": { "type": "string" }, "description": "Scope to one swarmail inbox" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "start", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": { "200": { "description": "Mailpit MessageSummary envelope" }, "400": { "description": "Missing query or unknown kind" } }
      }
    },
    "/api/v2/messages": {
      "get": {
        "summary": "MailHog v2 list shape over the whole store",
        "description": "{total, count, start, items} where items are MailHog data.Message objects (Content.{Headers,Body,Size,MIME}, Raw.{From,To,Data,Helo}). `?inbox=` scopes; `?limit=`/`?start=` page.",
        "parameters": [
          { "name": "inbox", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "start", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": { "200": { "description": "MailHog v2 envelope" } }
      }
    },
    "/api/v2/search": {
      "get": {
        "summary": "MailHog v2 search",
        "parameters": [
          { "name": "query", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "kind", "in": "query", "schema": { "type": "string", "enum": ["from", "to", "containing"], "default": "containing" } },
          { "name": "inbox", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "MailHog v2 envelope" }, "400": { "description": "Bare 400 (no body) for an unknown kind or empty query, as upstream answers" } }
      }
    },
    "/api/v1/delete-all": {
      "delete": { "summary": "MailHog's wipe alias", "responses": { "200": { "description": "{\"removed\": n}" } } }
    },
    "/api/v1/messages/{id}": {
      "get": {
        "summary": "Get one email in full (bodies, links, codes)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Email" } } } }, "404": { "description": "Not found" } }
      },
      "delete": {
        "summary": "Delete one email",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not found" } }
      }
    },
    "/api/v1/messages/{id}/raw": {
      "get": {
        "summary": "Raw RFC 5322 source",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "text/plain" }, "404": { "description": "Not found" } }
      }
    },
    "/api/v1/chaos": {
      "get": { "summary": "Current chaos rules", "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChaosConfig" } } } } } },
      "put": {
        "summary": "Replace chaos rules (event: connect|mail_from|rcpt|data; probability 0-100; optional error line and delay_ms)",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChaosConfig" } } } },
        "responses": { "200": { "description": "OK" } }
      },
      "delete": { "summary": "Disable all chaos", "responses": { "200": { "description": "OK" } } }
    },
    "/api/v1/webhooks": {
      "get": { "summary": "Current webhook targets", "responses": { "200": { "description": "OK" } } },
      "put": {
        "summary": "Replace webhook targets. Each delivery POSTs {event:\"received\", email:{...}} with X-Swarmail-Secret; retried 4x (100ms..1.6s). http:// and https:// targets (https over TLS via reqwest + rustls, verification on; optional per-target ca_pem).",
        "requestBody": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/WebhookTarget" } } } } },
        "responses": { "200": { "description": "OK" } }
      },
      "delete": { "summary": "Remove all webhook targets", "responses": { "200": { "description": "OK" } } }
    }
  },
  "components": {
    "schemas": {
      "EmailAddress": {
        "type": "object",
        "properties": { "name": { "type": ["string", "null"] }, "address": { "type": "string" } }
      },
      "Email": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "inbox": { "type": "string" },
          "from": { "$ref": "#/components/schemas/EmailAddress" },
          "to": { "type": "array", "items": { "$ref": "#/components/schemas/EmailAddress" } },
          "cc": { "type": "array", "items": { "$ref": "#/components/schemas/EmailAddress" } },
          "recipients": { "type": "array", "items": { "type": "string" } },
          "subject": { "type": ["string", "null"] },
          "received_at": { "type": "string" },
          "received_ms": { "type": "integer" },
          "size": { "type": "integer" },
          "text": { "type": ["string", "null"] },
          "html": { "type": ["string", "null"] },
          "links": { "type": "array", "items": { "type": "string" } },
          "codes": { "type": "array", "items": { "type": "string" } },
          "message_id": { "type": "string", "description": "The Message-ID header value, angle brackets stripped; absent when the mail has none" },
          "in_reply_to": { "type": "string", "description": "The first In-Reply-To id; absent when the mail has none" },
          "references": { "type": "array", "items": { "type": "string" }, "description": "The References ids in header order; absent when the mail has none" },
          "session": { "type": "integer" }
        }
      },
      "Thread": {
        "type": "object",
        "properties": {
          "key": { "type": "string", "description": "Stable short key (16 hex chars) usable in URLs and the thread endpoint" },
          "subject": { "type": "string", "description": "The newest mail's display subject; absent when no mail in the thread has one" },
          "count": { "type": "integer" },
          "last_received_ms": { "type": "integer" },
          "emails": { "type": "array", "items": { "$ref": "#/components/schemas/Email" }, "description": "The conversation, oldest first" }
        }
      },
      "EmailList": {
        "type": "object",
        "properties": { "total": { "type": "integer" }, "emails": { "type": "array", "items": { "$ref": "#/components/schemas/Email" } } }
      },
      "AwaitResponse": {
        "type": "object",
        "properties": {
          "matched": { "type": "integer" },
          "total": { "type": "integer" },
          "emails": { "type": "array", "items": { "$ref": "#/components/schemas/Email" } }
        }
      },
      "InboxSummary": {
        "type": "object",
        "properties": { "name": { "type": "string" }, "count": { "type": "integer" } }
      },
      "SeedRequest": {
        "type": "object",
        "properties": {
          "from": { "type": "string" },
          "to": { "type": "string" },
          "subject": { "type": "string" },
          "text": { "type": "string" },
          "html": { "type": "string" }
        },
        "required": ["to"]
      },
      "ChaosRule": {
        "type": "object",
        "properties": {
          "probability": { "type": "integer" },
          "error": { "type": ["string", "null"] },
          "delay_ms": { "type": "integer" }
        }
      },
      "ChaosConfig": {
        "type": "object",
        "properties": {
          "connect": { "$ref": "#/components/schemas/ChaosRule" },
          "mail_from": { "$ref": "#/components/schemas/ChaosRule" },
          "rcpt": { "$ref": "#/components/schemas/ChaosRule" },
          "data": { "$ref": "#/components/schemas/ChaosRule" }
        }
      },
      "WebhookTarget": {
        "type": "object",
        "properties": {
          "url": { "type": "string", "description": "http:// or https://; https is delivered over TLS (reqwest + rustls, verification on)" },
          "inbox": { "type": ["string", "null"] },
          "secret": { "type": ["string", "null"] },
          "ca_pem": { "type": ["string", "null"], "description": "Optional PEM of an extra root CA to trust for this target, on top of the built-in root store" }
        }
      }
    }
  }
}